edit: turn CGwmSpatialWeight into template class
